services/network/HttpClientImpl: call correct stream overload for platforms where std::size_t is defined as unsigned long.

+++ services/network/HttpClientImpl.cpp
@@ -61,7 +61,7 @@ namespace services
     void HttpRequestFormatter::AddContentLength(std::size_t size)
     {
         infra::StringOutputStream contentLengthStream(contentLength);
-        contentLengthStream << size;
+        contentLengthStream << static_cast<uint64_t>(size);
         contentLengthHeader.Emplace("content-length", contentLength);
     }
